Files Uploaded
Many columns within the dataset exhibit noise yet hold significant information. Employing feature engineering can yield improved outcomes. The challenge lies in the dataset's limited size; nevertheless, you anticipate achieving commendable accuracy through skillful techniques. While ample data is advantageous, the potency of thoughtful feature engineering can counterbalance the data shortage. The endeavor involves creating a website to estimate laptop prices based on user configurations

Below are the Data Dictionaries present in dataset
Column Name 	Description
Company 	The brand or company that manufactures the laptop.
TypeName 	The general type or category of the laptop.
Inches 	The size of the laptop screen in inches.
ScreenResolution 	The display resolution of the laptop screen.
Cpu 	The central processing unit (processor) of the laptop.
Ram 	The amount of random-access memory (RAM) in the laptop.
Memory 	The storage capacity of the laptop's memory.
Gpu 	The graphics processing unit (GPU) of the laptop.
OpSys 	The operating system installed on the laptop.
Weight 	The weight of the laptop.
Price 	The price of the laptop.